BAC-264: KMS Integration (#44)
* kms

* envelope encryption

* scheduler tee dockerfile

* new image

* fixes and server Dockerfile

* use kmstool cli

* bake in public key

* debugging

* inlcude kmstool cli in docker image

* wrong file debugging

* heavy img

* pub key from parent

* debugging

* ciphertext prefix

* no prefix

* Added support for AWS STS

* Added missing file

* Fix

* Retrieve creds from EC2 metadata

* Fix

* Fix

* Fix

* use pem file

* fix

* Fix

* Fix

* Revert pem file

* fix

* Fix

* Fix

* README

* fmt

* Merge conflicts

* Review part 1.

* Review

* Review review

* Review

* Update crates/shielder-scheduler-tee/README.md


* Fixes

* Docs

* Fix

* Fix

* Nits

* Fix build

* Update crates/shielder-scheduler-server/src/command_line_args.rs


* Review

* fix

* Review

* Review

* Fmt

---------

+++ tee/crates/shielder-prover-common/src/protocol.rs
@@ -5,7 +5,7 @@ use crate::{
     vsock::{VsockClient, VsockServer},
 };
 
-pub const VSOCK_PORT: u16 = 5000;
+pub const VSOCK_PORT: u32 = 5000;
 
 /// Padding constant for decrypting incoming payloads.
 /// The value is chosen to be large enough to accommodate typical payloads while also being efficient for transmission.
