+++ tee/crates/shielder-prover-server/src/command_line_args.rs
@@ -13,7 +13,7 @@ pub struct CommandLineArgs {
     /// Internal port on which host and tee applications talks to each other
     /// This is the part of the vsock endpoint, which is tee_cid:tee_port
     #[arg(short, long, default_value_t = shielder_prover_common::protocol::VSOCK_PORT, env = "TEE_PORT")]
-    pub tee_port: u16,
+    pub tee_port: u32,
 
     /// Local IPv4 address on which this server listens to incoming HTTP connections
     #[arg(short, long, default_value = "0.0.0.0", env = "BIND_ADDRESS")]
