+++ tee/crates/shielder-prover-server/src/handlers/mod.rs
@@ -17,7 +17,7 @@ pub mod metrics;
 pub mod tee_public_key;
 
 async fn request(state: Arc<AppState>, request: Request) -> Result<Json<Response>, VsockError> {
-    let mut tee_client = ProverClient::new(state.options.tee_cid, state.options.tee_port as u32)
+    let mut tee_client = ProverClient::new(state.options.tee_cid, state.options.tee_port)
         .instrument(info_span!(BuildingVsocksConnection.name()))
         .await?;
